Удаление фотографии беседы и фиксы

+++ app/src/main/java/ru/melodin/fast/adapter/UserAdapter.kt
@@ -66,8 +66,8 @@ class UserAdapter : RecyclerAdapter<VKUser, UserAdapter.ViewHolder> {
         if (ArrayUtil.isEmpty(data)) return
 
         when (data[0] as String) {
-            Keys.KEY_USER_OFFLINE -> setUserOnline(false, data[1] as Int, data[2] as Long)
-            Keys.KEY_USER_ONLINE -> setUserOnline(true, data[1] as Int, data[2] as Long)
+            Keys.KEY_USER_OFFLINE -> setUserOnline(false, data[1] as Int, data[2] as Int)
+            Keys.KEY_USER_ONLINE -> setUserOnline(true, data[1] as Int, data[2] as Int)
             Keys.CONNECTED -> {
                 fragment ?: return
                 if (fragment!!.isLoading)
@@ -84,12 +84,12 @@ class UserAdapter : RecyclerAdapter<VKUser, UserAdapter.ViewHolder> {
         return -1
     }
 
-    private fun setUserOnline(online: Boolean, userId: Int, time: Long) {
+    private fun setUserOnline(online: Boolean, userId: Int, time: Int) {
         for (i in 0 until itemCount) {
             val user = getItem(i)
             if (user.id == userId) {
                 user.isOnline = online
-                user.lastSeen = time
+                user.lastSeen = time.toLong()
                 notifyItemChanged(i, -1)
             }
         }
